requirements.txt
fastapi[all]
uvicorn==0.27.1
Flask==2.2.0
gunicorn
Werkzeug==2.2.0
line-bot-sdk==2.4.3
requests==2.31.0
groq
taiwanlottery
beautifulsoup4
yfinance
pandas
numpy
openai>=1.3.8
lxml
pypinyin
pyvi